+++ utils/src/match-screenshot.ts
@@ -1,0 +1,92 @@
+const PREFIX_DIFFERENTIATOR = '___';
+const SUFFIX_TEST_IDENTIFIER = '.spec.ts';
+const SCREENSHOTS_FOLDER_NAME = 'screenshots';
+
+function getTestFolderPathFromScripts(rawName?: string) {
+  const relativeTestPath = Cypress.spec.relative;
+
+  if (!relativeTestPath) {
+    throw new Error('❌ Could not find matching script in the Cypress DOM to infer the test folder path');
+  }
+
+  const testName = relativeTestPath.substring(relativeTestPath.lastIndexOf('/') + 1, relativeTestPath.lastIndexOf(SUFFIX_TEST_IDENTIFIER));
+  const name = rawName || testName;
+
+  const screenshotsFolder = `${SCREENSHOTS_FOLDER_NAME}/${relativeTestPath.substring(0, relativeTestPath.lastIndexOf(testName))}${name}`;
+
+  return {
+    name,
+    screenshotsFolder
+  };
+}
+
+function verifyImages() {
+  if (Cypress.$('img:visible').length > 0) {
+    cy.document()
+      .its('body')
+      .find('img')
+      .filter(':visible')
+      .then(images => {
+        if (images) {
+          cy.wrap(images).each($img => {
+            cy.wrap($img).should('exist').and('have.prop', 'naturalWidth');
+          });
+        }
+      });
+  }
+}
+
+function forceFont() {
+  const iframe = window.parent.document.querySelector('iframe');
+  const contentDocument = iframe?.contentDocument;
+
+  if (contentDocument) {
+    const style = contentDocument.createElement('style');
+    style.type = 'text/css';
+    style.appendChild(contentDocument.createTextNode('* { font-family: Arial !important; }'));
+    contentDocument.head.appendChild(style);
+  }
+}
+
+export type MatchScreenshotArgs = {
+  rawName?: string;
+  options?: Partial<Cypress.ScreenshotOptions>;
+};
+
+export function matchScreenshot(subject: Cypress.JQueryWithSelector | Window | Document | void, args?: MatchScreenshotArgs) {
+  const { rawName, options = {} } = args ?? {};
+  forceFont();
+  verifyImages();
+
+  const { name, screenshotsFolder } = getTestFolderPathFromScripts(rawName);
+
+  cy.task('baseExists', screenshotsFolder).then(hasBase => {
+    const type = 'new';
+    const target = subject ? cy.wrap(subject) : cy;
+    // For easy slicing of path ignoring the root screenshot folder
+    target.screenshot(`${PREFIX_DIFFERENTIATOR}${screenshotsFolder}/${type}`, options);
+
+    if (!hasBase) {
+      cy.task('createNewScreenshot', screenshotsFolder).then(() => {
+        cy.task('log', `❌A new base image was created for ${name}. Create this as a new base image via Comparadise!`);
+      });
+    }
+
+    cy.task('compareScreenshots', screenshotsFolder).then(diffPixels => {
+      if (diffPixels === 0) {
+        cy.log(`✅Actual image of ${name} was the same as base`);
+
+        return null;
+      }
+
+      const screenshotUrl = Cypress.env('BUILD_URL') ? `${Cypress.env('BUILD_URL')}artifact/${screenshotsFolder}` : screenshotsFolder;
+
+      throw new Error(
+        `❌Actual image of ${name} differed by ${diffPixels} pixels.
+             See the diff image for more details >>  ${screenshotUrl}/diff.png`
+      );
+    });
+
+    return null;
+  });
+}
