+++ utils/src/screenshots.ts
@@ -3,11 +3,6 @@ import { PNG } from 'pngjs';
 import { getDiffPixels } from './images';
 import { createImageFileName } from './files';
 
-/**
- * Checks if a base image exists
- * @param path - Folder path where you can find the base.png image
- * @returns true if path/base.png exists, false if not.
- */
 export function baseExists(path: string) {
   const fileName = createImageFileName(path, 'base');
   const exists = fs.existsSync(fileName);
@@ -29,11 +24,6 @@ export function createNewScreenshot(screenshotFolder: string) {
   return null;
 }
 
-/**
- * Runs a visual regression test.
- * @param screenshotFolder - Full screenshots folder where the base/new/diff
- *                           images will be compared and written to.
- */
 export function compareScreenshots(screenshotFolder: string) {
   const basePath = createImageFileName(screenshotFolder, 'base');
   const actualPath = createImageFileName(screenshotFolder, 'new');
@@ -56,49 +46,4 @@ export function compareScreenshots(screenshotFolder: string) {
   }
 
   return diffPixels;
-}
-
-/**
- * Renames all root cypress screenshots to where the test was actually run.
- * Should NOT be used standalone. Works with the matchScreenshot task.
- * @param {Cypress.ScreenshotDetails} details
- */
-export function onAfterScreenshot(details: Cypress.ScreenshotDetails): Promise<Cypress.AfterScreenshotReturnObject> {
-  console.log('🧸 Screenshot was saved to:', details.path);
-  if (!details.path.match('cypress')) {
-    return Promise.resolve({});
-  }
-
-  const getNewPath = (path: string) => {
-    let newPath = path.slice(path.lastIndexOf('___') + 3);
-    console.log(newPath);
-
-    if (newPath.startsWith('/')) {
-      newPath = `.${newPath}`;
-    }
-
-    return newPath;
-  };
-
-  const newPath = getNewPath(details.path);
-  const newPathDir = newPath.substring(0, newPath.lastIndexOf('/'));
-
-  try {
-    fs.mkdirSync(newPathDir, { recursive: true });
-    console.log('🧸 No screenshot folder found in the package. Created new screenshot folder:', newPathDir);
-  } catch (err) {
-    console.error('❌ Error creating new screenshot folder:', newPathDir, err);
-  }
-
-  return new Promise((resolve, reject) => {
-    fs.rename(details.path, newPath, err => {
-      if (err) {
-        reject(err);
-      }
-
-      // because we renamed/moved the image, resolve with the new path
-      // so it is accurate in the test results
-      resolve({ path: newPath });
-    });
-  });
 }
