+++ utils/src/visual-test-setup.ts
@@ -1,0 +1,17 @@
+import { onAfterScreenshot } from './on-after-screenshot';
+import { baseExists, compareScreenshots, createNewScreenshot } from './screenshots';
+
+export function visualTestSetup(on: Cypress.PluginEvents, config: Cypress.PluginConfigOptions) {
+  on('after:screenshot', onAfterScreenshot);
+  on('task', {
+    baseExists,
+    compareScreenshots,
+    createNewScreenshot,
+    log: (message: string) => {
+      console.log(message);
+      return null;
+    }
+  });
+
+  return config;
+}
